+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.6.5 (2025-05-23)
 - support "snapshot:" as well as "resolver:"
 - add -S,--make-all-lts flag (pbrisbin,#4)
+- error for lts11 with stack > 3
 - add 9.8 alias for lts23
 
 ## 0.6.4 (2024-10-13)